+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 
 * PEP8 standards
 * Inserting dictionary object with keys as column name to database
+* Optimized for pydoc