+++ CHANGELOG.md
@@ -16,3 +16,16 @@
 * Inserting dictionary object with keys as column name to database
 * Optimized for pydoc
 * Added method to update table
+
+** v 1.0 **:
+
+* Added to python modules
+* Docs added
+
+** v 1.2 **:
+
+* Added commit boolean to commit all messages in one
+
+** v 1.3 **:
+
+* Added update on key exists to update some columns if key is duplicate